+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * Added button to toggle keyframes in the property view
     * You can now click-drag to edit properties
     * Made the keyframe status display in the timeline more consistent
+    * Added context menus to the properties dock
 * I/O:
     * Improved matte import and export for lottie
     * Improved mask properties import for lottie